pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "energy-grid-balancer"
version = "1.0.0"
description = "Real-world renewable energy grid management environment for OpenEnv / RL training."
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11"
keywords = ["openenv", "reinforcement-learning", "energy", "grid", "environment"]

dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.5.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "openai>=1.30.1",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]
